accc: fix some visual studio warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@7716 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: c5d98b4073908ec0d675c710c23b1c2c0c62b467
Former-commit-id: b9cba888c8f25c8f997f93805f4ca3548c7c43cc

+++ src/libs/portability/portability.cpp
@@ -24,7 +24,7 @@
 // Code provided by Thierry Thomas.
 #ifndef HAVE_STRNDUP
 
-char *strndup(const char *str, int len)
+char *strndup(const char *str, size_t len)
 {
 	if (!str || len < 0)
 		return 0;
